:recycle: Refactor: 로직분리
